LayoutTests/platform/chromium/inspector/debugger/function-location-expected.txt

Tests that Debugger.getFunctionLocation command returns location. Bug 71808

Debugger was enabled.

Running: testGetFirstLineFunctionLocation
firstLineFunction type = function
didGetLocation: 
lineNumber: 5
columnNumber: 36
scriptId is valid: true

Running: testGetNonFirstLineFunctionLocation
nonFirstLineFunctiontype = function
didGetLocation: 
lineNumber: 10
columnNumber: 32
scriptId is valid: true
Debugger was disabled.
